fixing initialize in Counter
p. 48 "However, as we mentioned, the initialize message is sent to the newly created instance." - we do not mention abot initialize message yet. It would be nice to explain the initialization process.
Comment: "set the initial value of the VALUE to 0" sould be "set the initial value of the count to 0"