revise update vs refresh in inspector to address
p. 61 "To refresh the values, click on the update button (the blue little circle)..." - There are not "update" button in the Inspector's window. It is Refresh button, but we do not need to click it because the value updates automaticly. By the way, the using of rotate handle of the cell's morphic halo frustrates the Inspector